+++ ui/components/multichain/account-details/account-details-display.js
@@ -14,7 +14,7 @@ import {
 } from '../../../selectors';
 import { isAbleToExportAccount } from '../../../helpers/utils/util';
 import {
-  BUTTON_SECONDARY_SIZES,
+  ButtonSecondarySize,
   ButtonSecondary,
   Box,
 } from '../../component-library';
@@ -78,7 +78,7 @@ export const AccountDetailsDisplay = ({
       {exportPrivateKeyFeatureEnabled ? (
         <ButtonSecondary
           block
-          size={BUTTON_SECONDARY_SIZES.LG}
+          size={ButtonSecondarySize.Lg}
           variant={TextVariant.bodyMd}
           onClick={() => {
             trackEvent({