+++ ui/components/multichain/import-account/import-account.js
@@ -82,7 +82,7 @@ export const ImportAccount = ({ onActionComplete }) => {
   }
 
   function getLoadingMessage(strategy) {
-    if (strategy === 'JSON File') {
+    if (strategy === 'json') {
       return (
         <>
           <Text width={BlockSize.ThreeFourths} fontWeight={FontWeight.Bold}>